refactor(backend): extract AttentionRegistry from AgentService
Everything an agent can be blocked on waiting for the user — a question, an
approval, a browser takeover — moves out of the god class into
`src/backend/agent/attention-registry.ts`. The three pending maps and the
fourteen surface/respond/clear methods go with it; `respondToPrompt`,
`respondToApproval` and `respondToBrowserTakeover` stay on the facade as
one-line delegations, so the public surface is byte-identical.

The three belong together because the product already treats them as one
queue: the runtime snapshot budgets them against a single attention limit in
a fixed order, duplication refuses while any is outstanding, and a turn
ending clears all three at once. Each also owes the provider a response, so
every path out of these maps either answers or cancels — an entry silently
dropped is an agent stuck forever.

Two ports keep the registry from learning the neighbouring domains:

- `RoutineAttention` (`markNeedsAttention` / `markRunningForTurn`) — all
  seven `surface*` methods raise a routine run's attention flag and all three
  `respondTo*` lower it. Defined here and implemented inline by the facade;
  `RoutineScheduler` takes it over in the next PR.
- `HostedSiteApprovals` (`prepareApproval` / `resolveApproval`) —
  `HostedSiteCoordinator` satisfies it structurally, so the registry never
  learns what a site is.

Per the plan, the constructor's `browser.onChanged` listener no longer
iterates the takeover map by hand: a tab closing calls
`cancelTakeoversForMissingTabs(tabs)`, which keeps the "a takeover whose tab
disappeared can never be answered" rule inside the registry that owns it.

`runtimeAttention()` returns the four snapshot fields as one record spread
into `fitRuntimeSnapshot`, so the ordering of the attention budget — prompts
first, takeovers last — cannot drift from the state it budgets.

Verification, test files untouched: all six agent-service shards and the four
`src/backend/agent/*.test.ts` files pass (93 tests), `bun run typecheck:node`
is clean — `Pick<AgentService, …>` in `team-api-server.ts` is the detector
for an accidental surface change — and the private-member delta against the
pre-edit file is exactly the seventeen removed names plus `#attention`.
`BEGIN IMMEDIATE` is still one occurrence under `src/backend/agent*`.

Delta: −486 in `agent-service.ts` (3626 → 3140), +619 in the new file,
against the plan's −380 / +410; the overshoot on both sides is the fourteen
method bodies plus the two port interfaces and their doc comments.
